+++ tests/testthat/test-get_objects.R
@@ -41,3 +41,13 @@ test_that("get_objects() warns for batch sizes above 50", {
 test_that("get_objects() can fallback on dplyr if data.table isn't installed", {
   # Implement with mocked binding or with withr::with_libpath ?
 })
+
+test_that("get_objects() returns POSIXct dates and not time since 1970",{
+  # fetch 150 random records
+  rato_obs <-
+    get_objects(object_ids = sample(list_object_ids(), size = 150))
+  
+  dplyr::select(rato_obs, dplyr::contains("Datum")) %>% 
+    purrr::walk(expect_s3_class, "POSIXct")
+  
+})