+++ NEWS.md
@@ -1,3 +1,2 @@
 # ratatouille (development version)
-
-
+- Add `ratatouille()`, a function to fetch all available invasive species data from RATO in a single function call. (#36)